mcp/requirements.txt
mcp[sse]>=1.0.0
elasticsearch>=8.0.0
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
